swim_arrival_processor/requirements.txt
mysql-connector
redis
arrow>=1
